Refactor and flatten the main menu components.
